Extract HeterogeneousAdapter SectionedAdapter implementation into another class
This allows creating fast-scrolling lists with Sections that do not implement the SectionedAdapter interface (the Album page is a particular example of this)
